Move scope analyzer to sub module.
