+++ src/analyzer/mod.rs
@@ -1,6 +1,3 @@
-use crate::ast::statement::Stmt;
-use std::fmt;
-
 #[cfg(test)]
 mod tests;
 
@@ -11,35 +8,4 @@ pub trait SemanticAnalyzer<A, B> {
     fn analyze(&self, input: A) -> Result<B, Self::Error>;
 }
 
-#[derive(PartialEq, Debug)]
-pub enum ScopeAnalyzerErr {
-    Unspecified,
-}
-
-impl fmt::Display for ScopeAnalyzerErr {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Unspecified => write!(f, "unspecified resolver error"),
-        }
-    }
-}
-
-pub struct ScopeAnalyzer {}
-
-impl SemanticAnalyzer<Vec<Stmt>, Vec<Stmt>> for ScopeAnalyzer {
-    type Error = ScopeAnalyzerErr;
-
-    fn analyze(&self, input: Vec<Stmt>) -> Result<Vec<Stmt>, Self::Error> {
-        input.into_iter().map(|stmt| self.analyze(stmt)).collect()
-    }
-}
-
-impl SemanticAnalyzer<Stmt, Stmt> for ScopeAnalyzer {
-    type Error = ScopeAnalyzerErr;
-
-    fn analyze(&self, input: Stmt) -> Result<Stmt, Self::Error> {
-        match input {
-            _ => Err(ScopeAnalyzerErr::Unspecified),
-        }
-    }
-}
+pub mod scope;
